Improve argument parser user experience
See https://github.com/probonopd/linuxdeployqt/issues/256#issuecomment-373916136

The argument parser requires flags expecting a payload to be passed in
the form -arg=payload. If not, it fails to recognize them and falsely
reports "argument must not start with --". These usability issues
should be gone with these two minor changes.